Refactor MoveAnimal and add documentation to the methods and its' implementations. Refactor AnimalStock name to AnimalBaseStock.
