+++ src/main/java/hu/eszter/bokkon/model/participants/MoveAnimal.java
@@ -2,15 +2,19 @@ package hu.eszter.bokkon.model.participants;
 
 import hu.eszter.bokkon.model.animal.Animal;
 
-import java.util.List;
-
 public interface MoveAnimal {
 
-    void addAnimal(Animal newAnimal);
+    /**
+     * Adds given number of the same type of animal to an animal stock
+     * @param animal type of animal to be added
+     * @param howMany number of the same type of animal to be added
+     */
+    void addAnimals(Animal animal, int howMany);
 
-    void addAnimals(List<Animal> newAnimals);
-
-    void removeAnimal(Animal newAnimal);
-
-    void removeAnimals(List<Animal> newAnimals);
+    /**
+     * Removes given number of the same type of animal from an animal stock
+     * @param animal type of animal to be removed
+     * @param howMany number of the same type of animal to be removed
+     */
+    void removeAnimals(Animal animal, int howMany);
 }
